feat: add weekly payroll calculations
